Check that a client is allowed to send netobjects before changing any state
Add a static AcceptClientSyncStatic() method to nNetObject classes to
report that permission to the system before any objects need to be created.

Previously, that was the sole responsibility of the virtual AcceptClientSync(),
which required that we allow the object the client sends to be created, which
already can cause damage. Furthermore, the cleanup in the case where creation
should not be allowed was faulty and crashed.