Remove DATE_R
We always use either SOURCE_DATE_EPOCH as date source or the
current date. Before the reproducibility patch, we would prefer
to use the last change of the source file, which requires
the -r flag, which we therefore tested for. Still supporting
that now would require too much if/else logic.